[Feat] finish NewBPay form action
