+++ .env.example
@@ -1,0 +1,2 @@
+NEXT_PUBLIC_API_URL=https://pi.rocket-coding.com
+NEXT_PUBLIC_NEWBPAY_URL=https://ccore.newebpay.com/MPG/mpg_gateway
